refactor(nicknamer): inline `NickService` into `NicknamerImpl`
Inlined the `NickService` functionality into `NicknamerImpl`, simplifying the architecture by removing the separate `NickService` abstraction. Updated implementations to directly handle nickname operations within `NicknamerImpl`, improving code cohesion and maintainability.
