+++ nicknamer/src/nicknamer/mod.rs
@@ -7,7 +7,6 @@ use crate::nicknamer::connectors::discord;
 use async_trait::async_trait;
 use commands::Error;
 use commands::names::NamesRepository;
-use commands::nick::{NickService, NickServiceImpl};
 use commands::reveal::{Revealer, RevealerImpl};
 use connectors::discord::DiscordConnector;
 
@@ -34,6 +33,80 @@ impl<'a, REPO: NamesRepository, DISCORD: DiscordConnector> NicknamerImpl<'a, REP
             discord_connector,
         }
     }
+
+    async fn admonish_for_violating_party_guidelines(&self) -> Result<(), Error> {
+        let reply = "You dare to rename our great General Secretary??? Away with your impudence!";
+        self.discord_connector.send_reply(reply).await?;
+        Ok(())
+    }
+
+    async fn change_member_nick_name(
+        &self,
+        member: &discord::ServerMember,
+        new_nick_name: &str,
+    ) -> Result<(), Error> {
+        match self
+            .discord_connector
+            .change_member_nick_name(member.id, new_nick_name)
+            .await
+        {
+            Ok(()) => match &member.nick_name {
+                Some(nick_name) => {
+                    self.send_reply_for_member_with_nick_name(member, new_nick_name, nick_name)
+                        .await?
+                }
+                None => {
+                    self.send_reply_for_member_without_nick_name(member, new_nick_name)
+                        .await?
+                }
+            },
+            Err(err) => {
+                let reply = match err {
+                    discord::Error::NotEnoughPermissions => {
+                        let role_to_mention = self
+                            .discord_connector
+                            .get_role_by_name(config::CODE_MONKEYS_ROLE_NAME)
+                            .await?;
+                        format!(
+                            "Some devilry restricts my power. {} please investigate the rogue member {}",
+                            role_to_mention.mention(),
+                            member.mention
+                        )
+                    }
+                    err => format!("You fool! You messed it up!: {}", err),
+                };
+                self.discord_connector.send_reply(&reply).await?;
+            }
+        }
+        Ok(())
+    }
+
+    async fn send_reply_for_member_without_nick_name(
+        &self,
+        member: &discord::ServerMember,
+        new_nick_name: &str,
+    ) -> Result<(), Error> {
+        let reply = format!(
+            "{} has been christened with the name {}!",
+            member.user_name, new_nick_name
+        );
+        self.discord_connector.send_reply(&reply).await?;
+        Ok(())
+    }
+
+    async fn send_reply_for_member_with_nick_name(
+        &self,
+        member: &discord::ServerMember,
+        new_nick_name: &str,
+        nick_name: &String,
+    ) -> Result<(), Error> {
+        let reply = format!(
+            "Changed {}'s nickname from '{}' to '{}'",
+            member.user_name, nick_name, new_nick_name
+        );
+        self.discord_connector.send_reply(&reply).await?;
+        Ok(())
+    }
 }
 
 #[async_trait]
@@ -55,8 +128,13 @@ impl<'a, REPO: NamesRepository + Send + Sync, DISCORD: DiscordConnector + Send +
         member: &discord::ServerMember,
         new_nickname: &str,
     ) -> Result<(), Error> {
-        let nick_service = NickServiceImpl::new(self.discord_connector);
-        nick_service.nick(member, new_nickname).await
+        let owner_id = self.discord_connector.get_guild_owner_id().await?;
+        if member.id == owner_id {
+            self.admonish_for_violating_party_guidelines().await?;
+        } else {
+            self.change_member_nick_name(member, new_nickname).await?;
+        }
+        Ok(())
     }
 }
